Refactored services and files to get rid of unnecessary shared pointers; disabled google drive since it is impossible to test at the moment

+++ src/filesystem/fileaccesslocal.cpp
@@ -11,7 +11,7 @@ using namespace Files;
 Q_LOGGING_CATEGORY(gmFileAccessLocal, "gm.files.access.local")
 
 /// Read data from one file
-auto FileAccessLocal::getData(const QString &path, bool allowCache) -> std::shared_ptr<FileDataResult>
+auto FileAccessLocal::getData(const QString &path, bool allowCache) -> FileDataResult
 {
     Q_UNUSED(allowCache)
 
@@ -21,15 +21,15 @@ auto FileAccessLocal::getData(const QString &path, bool allowCache) -> std::shar
     {
         auto data = f.readAll();
         f.close();
-        return std::make_shared<FileDataResult>(data);
+        return FileDataResult(std::move(data));
     }
 
     qCWarning(gmFileAccessLocal()) << "Warning: Could not find file" << path << f.error() << f.errorString();
-    return std::make_shared<FileDataResult>(f.errorString());
+    return FileDataResult(f.errorString());
 }
 
 /// Read data from one file async
-auto FileAccessLocal::getDataAsync(const QString &path, bool allowCache) -> QFuture<std::shared_ptr<FileDataResult>>
+auto FileAccessLocal::getDataAsync(const QString &path, bool allowCache) -> QFuture<FileDataResult>
 {
     qCDebug(gmFileAccessLocal()) << "Getting data from file:" << path << "...";
 
@@ -37,13 +37,12 @@ auto FileAccessLocal::getDataAsync(const QString &path, bool allowCache) -> QFut
 }
 
 /// Read data from multiple files
-auto FileAccessLocal::getDataAsync(const QStringList &paths, bool allowCache)
-    -> QFuture<std::vector<std::shared_ptr<FileDataResult>>>
+auto FileAccessLocal::getDataAsync(const QStringList &paths, bool allowCache) -> QFuture<std::vector<FileDataResult>>
 {
     qCDebug(gmFileAccessLocal()) << "Getting data from multiple files:" << paths << "...";
 
     return QtConcurrent::run([paths, allowCache]() {
-        std::vector<std::shared_ptr<FileDataResult>> results;
+        std::vector<FileDataResult> results;
         results.reserve(paths.size());
 
         foreach (const auto &path, paths)
@@ -56,33 +55,33 @@ auto FileAccessLocal::getDataAsync(const QStringList &paths, bool allowCache)
 }
 
 /// Create a directory
-auto FileAccessLocal::createDir(const QDir &dir) -> std::shared_ptr<FileResult>
+auto FileAccessLocal::createDir(const QDir &dir) -> FileResult
 {
     if (dir.exists())
     {
         auto message = u"The directory %1 already exists ..."_s.arg(dir.path());
         qCWarning(gmFileAccessLocal()) << message;
-        return std::make_shared<FileResult>(message);
+        return FileResult(std::move(message));
     }
 
     if (!dir.mkpath(dir.path()))
     {
         auto message = u"Could not create directory %1."_s.arg(dir.path());
         qCWarning(gmFileAccessLocal()) << message;
-        return std::make_shared<FileResult>(message);
+        return FileResult(std::move(message));
     }
-    return std::make_shared<FileResult>(true);
+    return FileResult(true);
 }
 
 /// Create a directory
-auto FileAccessLocal::createDirAsync(const QString &path) -> QFuture<std::shared_ptr<FileResult>>
+auto FileAccessLocal::createDirAsync(const QString &path) -> QFuture<FileResult>
 {
     qCDebug(gmFileAccessLocal()) << "Creating directory:" << path << "...";
     return QtConcurrent::run(&FileAccessLocal::createDir, QDir(path));
 }
 
 /// Save data to a file
-auto FileAccessLocal::save(const QString &path, const QByteArray &data) -> std::shared_ptr<FileResult>
+auto FileAccessLocal::save(const QString &path, const QByteArray &data) -> FileResult
 {
     QFile f(path);
     QString errorMessage;
@@ -91,10 +90,10 @@ auto FileAccessLocal::save(const QString &path, const QByteArray &data) -> std::
     if (QFileInfo info(f); !info.dir().exists())
     {
         auto createDirResult = createDir(info.dir());
-        if (!createDirResult->success())
+        if (!createDirResult.success())
         {
-            errorMessage = u"Could not save file %1: %2"_s.arg(path, createDirResult->errorMessage());
-            return std::make_shared<FileResult>(false, errorMessage);
+            errorMessage = u"Could not save file %1: %2"_s.arg(path, createDirResult.errorMessage());
+            return FileResult(std::move(errorMessage));
         }
     }
 
@@ -121,17 +120,17 @@ auto FileAccessLocal::save(const QString &path, const QByteArray &data) -> std::
         qCWarning(gmFileAccessLocal()) << "Warning:" << errorMessage;
     }
 
-    return std::make_shared<FileResult>(status, errorMessage);
+    return FileResult(status, errorMessage);
 }
 
 /// Save data to a file async
-auto FileAccessLocal::saveAsync(const QString &path, const QByteArray &data) -> QFuture<std::shared_ptr<FileResult>>
+auto FileAccessLocal::saveAsync(const QString &path, const QByteArray &data) -> QFuture<FileResult>
 {
     qCDebug(gmFileAccessLocal()) << "Saving file:" << path << "...";
     return QtConcurrent::run(&FileAccessLocal::save, path, data);
 }
 
-auto FileAccessLocal::move(const QString &oldPath, const QString &newPath) -> std::shared_ptr<FileResult>
+auto FileAccessLocal::move(const QString &oldPath, const QString &newPath) -> FileResult
 {
     QFile f(oldPath);
     QFileInfo info(newPath);
@@ -139,11 +138,11 @@ auto FileAccessLocal::move(const QString &oldPath, const QString &newPath) -> st
     if (!info.dir().exists())
     {
         auto createDirResult = createDir(info.dir());
-        if (!createDirResult->success())
+        if (!createDirResult.success())
         {
-            auto errorMessage = u"Could not move %1 to %2: %3"_s.arg(oldPath, newPath, createDirResult->errorMessage());
+            auto errorMessage = u"Could not move %1 to %2: %3"_s.arg(oldPath, newPath, createDirResult.errorMessage());
             qCWarning(gmFileAccessLocal()) << "Warning:" << errorMessage;
-            return std::make_shared<FileResult>(false, errorMessage);
+            return FileResult(false, errorMessage);
         }
     }
 
@@ -152,59 +151,58 @@ auto FileAccessLocal::move(const QString &oldPath, const QString &newPath) -> st
         auto errorMessage =
             u"Could not move %1 to %2: %3 %4"_s.arg(oldPath, newPath, QString::number(f.error()), f.errorString());
         qCWarning(gmFileAccessLocal()) << "Warning:" << errorMessage;
-        return std::make_shared<FileResult>(false, errorMessage);
+        return FileResult(false, errorMessage);
     }
 
-    return std::make_shared<FileResult>(true);
+    return FileResult(true);
 }
 
 /// Move file from oldPath to newPath
-auto FileAccessLocal::moveAsync(const QString &oldPath, const QString &newPath) -> QFuture<std::shared_ptr<FileResult>>
+auto FileAccessLocal::moveAsync(const QString &oldPath, const QString &newPath) -> QFuture<FileResult>
 {
     qCDebug(gmFileAccessLocal()) << "Moving file:" << oldPath << "->" << newPath << "...";
     return QtConcurrent::run(&FileAccessLocal::move, oldPath, newPath);
 }
 
 /// Delete the file or folder at path
-auto FileAccessLocal::deleteAsync(const QString &path) -> QFuture<std::shared_ptr<FileResult>>
+auto FileAccessLocal::deleteAsync(const QString &path) -> QFuture<FileResult>
 {
     qCDebug(gmFileAccessLocal()) << "Deleting file:" << path << "...";
     return QtConcurrent::run([path]() {
         QFileInfo info(path);
 
-        if (!info.exists())
-            return std::make_shared<FileResult>(false, u"The file or folder %1 does not exist"_s.arg(path));
+        if (!info.exists()) return FileResult(false, u"The file or folder %1 does not exist"_s.arg(path));
 
         if (QDir dir(path); info.isDir() && !dir.removeRecursively())
         {
             auto errorMessage = u"Could not delete folder %1"_s.arg(path);
             qCWarning(gmFileAccessLocal()) << "Warning:" << errorMessage;
-            return std::make_shared<FileResult>(false, errorMessage);
+            return FileResult(false, errorMessage);
         }
         else if (QFile f(path); info.isFile() && !f.remove())
         {
             auto errorMessage =
                 u"Could not delete file %1: %2 %3"_s.arg(path, QString::number(f.error()), f.errorString());
             qCWarning(gmFileAccessLocal()) << "Warning:" << errorMessage;
-            return std::make_shared<FileResult>(false, errorMessage);
+            return FileResult(false, errorMessage);
         }
 
-        return std::make_shared<FileResult>(true);
+        return FileResult(true);
     });
 }
 
-auto FileAccessLocal::copy(const QString &path, const QString &copy) -> std::shared_ptr<FileResult>
+auto FileAccessLocal::copy(const QString &path, const QString &copy) -> FileResult
 {
     QFile f(path);
 
     if (QFileInfo info(copy); !info.dir().exists())
     {
         auto createDirResult = createDir(info.dir());
-        if (!createDirResult->success())
+        if (!createDirResult.success())
         {
-            auto errorMessage = u"Could not copy %1 to %2: %3"_s.arg(path, copy, createDirResult->errorMessage());
+            auto errorMessage = u"Could not copy %1 to %2: %3"_s.arg(path, copy, createDirResult.errorMessage());
             qCWarning(gmFileAccessLocal()) << "Warning:" << errorMessage;
-            return std::make_shared<FileResult>(false, errorMessage);
+            return FileResult(false, errorMessage);
         }
     }
 
@@ -213,54 +211,52 @@ auto FileAccessLocal::copy(const QString &path, const QString &copy) -> std::sha
         auto errorMessage =
             u"Could not copy %1 to %2: %3 %4"_s.arg(path, copy, QString::number(f.error()), f.errorString());
         qCWarning(gmFileAccessLocal()) << "Warning:" << errorMessage;
-        return std::make_shared<FileResult>(false, errorMessage);
+        return FileResult(false, errorMessage);
     }
 
-    return std::make_shared<FileResult>(true);
+    return FileResult(true);
 }
 
 /// Copy a file at path to a new path
-auto FileAccessLocal::copyAsync(const QString &path, const QString &copy) -> QFuture<std::shared_ptr<FileResult>>
+auto FileAccessLocal::copyAsync(const QString &path, const QString &copy) -> QFuture<FileResult>
 {
     qCDebug(gmFileAccessLocal()) << "Copying file:" << path << "->" << copy << "...";
     return QtConcurrent::run(&FileAccessLocal::copy, path, copy);
 }
 
 /// List files or/and folders in a directory
-auto FileAccessLocal::listAsync(const QString &path, bool files, bool folders)
-    -> QFuture<std::shared_ptr<FileListResult>>
+auto FileAccessLocal::listAsync(const QString &path, bool files, bool folders) -> QFuture<FileListResult>
 {
     return QtConcurrent::run([path, files, folders]() {
         QDir dir(path);
-        return std::make_shared<FileListResult>(path, dir.entryList(getDirFilter(false, folders)),
-                                                dir.entryList(getDirFilter(files, false)));
+        return FileListResult(path, dir.entryList(getDirFilter(false, folders)),
+                              dir.entryList(getDirFilter(files, false)));
     });
 }
 
 /// Check if a file exists
-auto FileAccessLocal::check(const QString &path, bool allowCache) -> std::shared_ptr<FileCheckResult>
+auto FileAccessLocal::check(const QString &path, bool allowCache) -> FileCheckResult
 {
     Q_UNUSED(allowCache)
 
     QFile f(path);
-    return std::make_shared<FileCheckResult>(path, f.exists());
+    return FileCheckResult(path, f.exists());
 }
 
 /// Check if a file exists async
-auto FileAccessLocal::checkAsync(const QString &path, bool allowCache) -> QFuture<std::shared_ptr<FileCheckResult>>
+auto FileAccessLocal::checkAsync(const QString &path, bool allowCache) -> QFuture<FileCheckResult>
 {
     return QtConcurrent::run(&FileAccessLocal::check, path, allowCache);
 }
 
 /// Check which files exist
-auto FileAccessLocal::checkAsync(const QStringList &paths, bool allowCache)
-    -> QFuture<std::shared_ptr<FileMultiCheckResult>>
+auto FileAccessLocal::checkAsync(const QStringList &paths, bool allowCache) -> QFuture<FileMultiCheckResult>
 {
     return QtConcurrent::run([paths, allowCache]() {
-        auto result = std::make_shared<FileMultiCheckResult>(true);
+        FileMultiCheckResult result(true);
         for (const auto &path : paths)
         {
-            result->add(check(path, allowCache));
+            result.add(check(path, allowCache));
         }
         return result;
     });
